AbyssalCraft 1.9.5
*Darklands Oak Wood Slabs, Darklands Oak Wood Fence and Dreadlands Wood Fence are now registered in the Ore Dictionary (closes #339)
*Added compatibility with Hardcore Darkness (if it's installed, all dimensions will be dark if the respective config options are enabled) (closes #191)
*Added a configuration category for options related to mod compatiblity
*Added configurable blacklists for Overworld structure and ore generation in other surface worlds (closes #334)
*Darkstone, Abyssal Stone etc, Abyssal Sand and Abyssal Sand Glass can now be manipulated through Chisel & Bits (closes #300)
*Textures for all Darkstone-based blocks, Darklands Oak Plank, Monolith Stone + Pillar, Shadow Fragment, Shadow Gem Shard and various Altar/Pedestal overlays have been revamped (courtesy of Cybercat5555)
*Be sure to name your anti-animals, or any anti-mob for that matter!
*Added a bunch of new spells (you can try them out by spawning in some scrolls)
*The game should no longer crash if a tamed and named horse dies without it's owner present in the same dimension
*Changed the damage cap on J'zahar to 20, and made it so that you can't bypass his invincibility frames
*Added a config option to toggle whether or not portals require a player to be nearby in order to rarely spawn mobs (closes #327)
*Portal blocks now have a spawn cap on how many of the rarely spawned mobs it can spawn (if enough are nearby the portal, no more spawn) (closes #327)
*Dread Spawns can no longer merge into a Greater Dread Spawn if 10 of those area within 32 blocks of them
*Greater Dread Spawns can no longer merge into a Lesser Dreadbeast of 4 of those are within 32 blocks of them, nor spawn Dread Spawns if 20 of those are within the same distance
*Lesser Dreadbeasts can't spawn Greater Dread Spawns if 10 of those are within 32 blocks of it, nor Dread Spawns of 20 of those are within the same distance
*The aforementioned spawn restrictions also applies for Cha'garoth when he spawns the same mobs during his fight (closes #290)
*Shoggoth Biomass now checks for Lesser Shoggoths within 32 blocks of it (instead of 16) when determining if it should spawn one
*Lesser Shoggoths can now be hurt through their main body again (instead of only their multiparts), so that blocks used to kill mobs in mob grinders hurt them again
*Rending Pedestals can now also target multipart entities
*Fixed a crash when a specific disruption was triggered by failing a ritual (closes #340)
*Items and Blocks with subtypes that can be used in a ritual now cycles through them when shown in the Necronomicon
*The Coralium and Dread enchantments can no longer be applied through an Anvil with an enchanted book
*Increased the follow range of the Skeleton Goliath
*Added doors for Darklands Oak and Dreadlands Wood (textures by Seth0067 and Cybercat5555) (closes #329)
*Fixed the models for the various cobblestone walls (now every connected wall block doesn't have the wall post part)
*Centered the MRE in its texture
*Added a texture with a snow overlay for the Dreadlands Grass
*Ritual Altars now check if the Necronomicon it's attempting to drain energy from during a ritual actually has energy stored before draining (in case someone has multiple books in their inventory)
*Lesser Shoggoths no longer spawn naturally in the Overworld (only through lairs)
*Added a configuration category for disabling things like Dreadlands spread (named "Wet Noodle") (closes #283, closes #286, closes #297)
*Updated the Lesser Shoggoth picture in the Necronomicon
*Added a config option to toggle whether or not boss dialogs display
*Added some variations to the Shoggoth Lairs, and statues (both normal and decorative) have a chance of generating inside them
*Changed the creature types of a few mobs (might add some more new ones to properly classify some mobs, or add some enchantments for extra damage against more stuff) (closes #278)
*Evil animals now always spawn like mobs, instead of broad daylight, like normal animals
*The disruptions that spawn random mobs from the current biome can no longer spawn Lesser Dreadbeasts
*Added Places of Power (multiblocks capable of harvesting PE from statues without causing Disruptions)
*Added a section with information about Places of Power in the Necronomicon (under Information in the Ritual section)
*There is currently only one Place of Power structure, but more will be added in future updates

+++ src/main/java/com/shinoow/abyssalcraft/common/entity/EntityAcidProjectile.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
  * AbyssalCraft
- * Copyright (c) 2012 - 2018 Shinoow.
+ * Copyright (c) 2012 - 2019 Shinoow.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the GNU Lesser Public License v3
  * which accompanies this distribution, and is available at
@@ -98,19 +98,20 @@ public class EntityAcidProjectile extends EntityThrowable {
 
 							double d = Math.sqrt(pos1.distanceSq(pos));
 
-							if(world.getGameRules().getBoolean("mobGriefing"))
-								for(int i = 0; i < d * pos.getDistance(pos1.getX(), pos1.getY(), pos1.getZ()); i++){
-									double i1 = i / pos.getDistance(pos1.getX(), pos1.getY(), pos1.getZ());
-									double xp = pos.getX() + vec.x * i1 + .5;
-									double yp = pos.getY() + vec.y * i1 + .5;
-									double zp = pos.getZ() + vec.z * i1 + .5;
-									BlockPos pos2 = new BlockPos(xp, yp, zp);
-									if(!world.isAirBlock(pos2) && world.getBlockState(pos2).getBlockHardness(world, pos2) < ACConfig.acidResistanceHardness && world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_ooze
-											&& world.getBlockState(pos2) != ACBlocks.stone.getDefaultState().withProperty(BlockACStone.TYPE, EnumStoneType.MONOLITH_STONE)
-											&& world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos2).getBlock().hasTileEntity(world.getBlockState(pos2))
-											&& world.getBlockState(pos2).getBlockHardness(world, pos2) != -1 && world.getBlockState(pos2).getBlock().canEntityDestroy(world.getBlockState(pos2), world, pos2, getThrower()))
-										world.destroyBlock(pos2, false);
-								}
+							if(!ACConfig.no_acid_breaking_blocks)
+								if(world.getGameRules().getBoolean("mobGriefing"))
+									for(int i = 0; i < d * pos.getDistance(pos1.getX(), pos1.getY(), pos1.getZ()); i++){
+										double i1 = i / pos.getDistance(pos1.getX(), pos1.getY(), pos1.getZ());
+										double xp = pos.getX() + vec.x * i1 + .5;
+										double yp = pos.getY() + vec.y * i1 + .5;
+										double zp = pos.getZ() + vec.z * i1 + .5;
+										BlockPos pos2 = new BlockPos(xp, yp, zp);
+										if(!world.isAirBlock(pos2) && world.getBlockState(pos2).getBlockHardness(world, pos2) < ACConfig.acidResistanceHardness && world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_ooze
+												&& world.getBlockState(pos2) != ACBlocks.stone.getDefaultState().withProperty(BlockACStone.TYPE, EnumStoneType.MONOLITH_STONE)
+												&& world.getBlockState(pos2).getBlock() != ACBlocks.shoggoth_biomass && !world.getBlockState(pos2).getBlock().hasTileEntity(world.getBlockState(pos2))
+												&& world.getBlockState(pos2).getBlockHardness(world, pos2) != -1 && world.getBlockState(pos2).getBlock().canEntityDestroy(world.getBlockState(pos2), world, pos2, getThrower()))
+											world.destroyBlock(pos2, false);
+									}
 						}
 					} else if(EntityUtil.damageShield(entity, damage*2))
 						entity.attackEntityFrom(AbyssalCraftAPI.acid, 1);
